+++ common/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_library(odr_common STATIC
         src/Constants.cpp
+        src/Html.cpp
         src/StringUtil.cpp
         src/TableCursor.cpp
         src/TablePosition.cpp
@@ -12,5 +13,7 @@ target_link_libraries(odr_common
         tinyxml2
 
         odr_access
+
+        odr-interface
         )
 set_property(TARGET odr_common PROPERTY POSITION_INDEPENDENT_CODE ON)
